+++ src/MixinError.js
@@ -1,0 +1,6 @@
+export default class MixinError extends Error {
+  constructor(args) {
+    super(args);
+    Error.captureStackTrace(this, MixinError);
+  }
+}