+++ src/alert.js
@@ -1,4 +1,4 @@
-import color from 'color';
+import color from '@bootstrap-styled/color';
 
 /**
  * @public