+++ src/background-variant.js
@@ -1,4 +1,4 @@
-import color from 'color';
+import color from '@bootstrap-styled/color';
 import { hoverFocus } from './hover';
 
 export const defaultProps = {
@@ -8,7 +8,7 @@ export const defaultProps = {
 /**
  * @public
  * @description Css helper for component backgrounds.
- * @param {bool} enableHoverMediaQuery Toggles hovering coloring if anchor is present.
+ * @param {boolean} enableHoverMediaQuery Toggles hovering coloring if anchor is present.
  * @param {string} selector
  * @param {string} bgColor
  * @returns {string}