+++ src/grid-framework.js
@@ -1,143 +1,102 @@
-import { makeGutters, makeCol, makeColModifier } from './grid';
-import { mediaBreakpointUp, breakpointInfix } from './breakpoints';
-
-export const defaultProps = {
-  '$grid-gutter-widths': {
-    xs: '30px',
-    sm: '30px',
-    md: '30px',
-    lg: '30px',
-    xl: '30px',
-  },
-  '$grid-breakpoints': {
-    xs: '0',
-    sm: '576px',
-    md: '768px',
-    lg: '992px',
-    xl: '1200px',
-  },
-  '$grid-columns': 12,
-  '$enable-grid-classes': true,
-};
-
 // Framework grid generation
 //
 // Used only by Bootstrap to generate the correct number of grid classes given
 // any value of `$grid-columns`.
 
-export function getGridColumn(gridGutterWidths = defaultProps['$grid-gutter-widths'], breakpoints = defaultProps['$grid-breakpoints']) {
-  return `
+import { makeCol, makeColOffset, defaultProps } from './grid';
+import { mediaBreakpointUp, breakpointInfix } from './breakpoints';
+
+/**
+ * @public
+ * @description Used only by Bootstrap to generate the correct number of grid classes given any value of `$grid-columns`.
+ * @param {boolean} [enableGridClasses=true] - if `false`, the function will return an empty string
+ * @param {number} [columns=12] - the number of columns
+ * @param {string|number} [gutter=30px] - the grid gutter width
+ * @param {object} [breakpoints={
+      xs: '0',
+      sm: '576px',
+      md: '768px',
+      lg: '992px',
+      xl: '1200px',
+    }] - The layout max-width grid breakpoint
+ *
+ * @returns {string} - css mixins utilities for bootstrap grid system
+ */
+export function makeGridColumns(
+  enableGridClasses = defaultProps['$enable-grid-classes'],
+  columns = defaultProps['$grid-columns'],
+  gutter = defaultProps['$grid-gutter-width'],
+  breakpoints = defaultProps['$grid-breakpoints']
+) {
+  // Common properties for all breakpoints
+  const gridColumn = `
     position: relative;
     width: 100%;
-    min-height: 1px; /* Prevent columns from collapsing when empty */
-    ${makeGutters(gridGutterWidths, breakpoints)}
+    min-height: 1px; // Prevent columns from collapsing when empty
+    padding-right: calc(${gutter} / 2);
+    padding-left: calc(${gutter} / 2);
   `;
-}
 
-export function getColumnGridColumn(gridColumns = defaultProps['$grid-columns'], gridBreakpoints = defaultProps['$grid-breakpoints'], gridGutterWidths = defaultProps['$grid-gutter-widths'], breakpoint) {
-  const columnList = [];
-  const infix = breakpointInfix(breakpoint, gridBreakpoints);
-  for (let i = 1; i <= gridColumns; i += 1) {
-    const column = `
-      &.col${infix}-${i},
-       & .col${infix}-${i}{
-        ${getGridColumn(gridGutterWidths, gridBreakpoints)}
+  return enableGridClasses ? Object.keys(breakpoints).map((bp) => {
+    const infix = breakpointInfix(bp, breakpoints);
+    // Allow columns to stretch full width below their breakpoints
+    return `${Array.from({ length: columns }, (_, ind) => {
+      const i = ind + 1;
+      return `&.col${infix}-${i}, & .col${infix}-${i} { ${gridColumn} }`;
+    }).join('\n')}
+      
+      &.col${infix},
+      & .col${infix}-auto {
+        ${gridColumn}
       }
-    `;
-    columnList.push(column);
-  }
-  return `
-    /* Allow columns to stretch full width below their breakpoints */
-    &.col${infix},
-     & .col${infix}{
-      ${getGridColumn(gridGutterWidths, gridBreakpoints)}
-    }
-
-    ${columnList.join('\n')}
-  `;
-}
-
-export function getMediaBreakpointUp(gridColumns = defaultProps['$grid-columns'], gridBreakpoints = defaultProps['$grid-breakpoints'], breakpoint) {
-  const infix = breakpointInfix(breakpoint, gridBreakpoints);
-  const basic = `
-    &.col${infix},
-     & .col${infix}{
-      flex-basis: 0;
-      flex-grow: 1;
-      max-width: 100%;
-    }
-    &.col${infix}-auto,
-     & .col${infix}-auto{
-      flex: 0 0 auto;
-      width: auto;
-    }
-  `;
-
-  const columnList = [];
-  for (let i = 1; i <= gridColumns; i += 1) {
-    const column = `
-      &.col${infix}-${i},
-       & .col${infix}-${i}{
-        ${makeCol(i, gridColumns)}
-      }
-    `;
-    columnList.push(column);
-  }
-  const modifierList = ['pull', 'push'];
-  const columnModifierList = [];
-  modifierList.forEach((modifier) => {
-    for (let i = 0; i <= gridColumns; i += 1) {
-      const columnModifier = `
-        &.${modifier}${infix}-${i},
-         & .${modifier}${infix}-${i}{
-          ${makeColModifier(modifier, i, gridColumns)}
+      
+${// Provide basic `.col-{bp}` classes for equal-width flexbox columns
+  mediaBreakpointUp(bp, breakpoints, `
+        &.col${infix},
+        & .col${infix} {
+          flex-basis: 0;
+          flex-grow: 1;
+          max-width: 100%;
         }
-      `;
-      columnModifierList.push(columnModifier);
-    }
-  });
-  // `$columns - 1` because offsetting by the width of an entire row isn't possible
-  const offsetColumnList = [];
-  for (let i = 0; i <= (gridColumns - 1); i += 1) {
-    if (infix !== 1 || i !== 0) { // Avoid emitting useless .offset-xs-0
-      const offsetColumn = `
-        &.offset${infix}-${i},
-         & .offset${infix}-${i}{
-          ${makeColModifier('offset', i, gridColumns)}
+        &.col${infix}-auto,
+        & .col${infix}-auto {
+          flex: 0 0 auto;
+          width: auto;
+          max-width: none; // reset earlier grid tiers
         }
-      `;
-      offsetColumnList.push(offsetColumn);
-    }
-  }
-  return mediaBreakpointUp(breakpoint, gridBreakpoints, `
-    ${basic}
-    ${columnList.join('\n')}
-    ${columnModifierList.join('\n')}
-    ${offsetColumnList.join('\n')}
-  `);
-}
-
-export function makeGridColumns(enableGridClasses = defaultProps['$enable-grid-classes'], gridColumns = defaultProps['$grid-columns'], gridGutterWidths = defaultProps['$grid-gutter-widths'], gridBreakpoints = defaultProps['$grid-breakpoints']) {
-  if (enableGridClasses) {
-    const gridColumnList = [];
-    Object.keys(gridBreakpoints).forEach((breakpoint) => {
-      const gridColumn = `
-        ${getColumnGridColumn(gridColumns, gridBreakpoints, gridGutterWidths, breakpoint)}
-        ${getMediaBreakpointUp(gridColumns, gridBreakpoints, breakpoint)}
-      `;
-      gridColumnList.push(gridColumn);
-    });
-    return `
-      ${gridColumnList.join('\n')}
-    `;
-  }
-  return '';
+        
+        ${Array.from({ length: columns }, (_, i) => `
+          &.col${infix}-${i + 1},
+          & .col${infix}-${i + 1} {
+            ${makeCol(i + 1, columns)}
+          }
+        `)}
+        
+        &.order${infix}-first,
+        & .order${infix}-first { order: -1; }
+        
+        &.order${infix}-last,
+        & .order${infix}-last { order: ${columns + 1}; }
+        
+        ${Array.from({ length: columns }, (_, i) => `
+          &.order${infix}-${i},
+          & .order${infix}-${i} { 
+            order: ${i}; 
+          }
+        `).join('\n')}
+        
+        ${// `$columns - 1` because offsetting by the width of an entire row isn't possible
+  Array.from({ length: columns - 1 }, (_, i) => `
+          &.offset${infix}-${i},
+          & .offset${infix}-${i} {
+            ${makeColOffset(i, columns)}
+          }
+        `).join('\n')}
+      `)}
+`;
+  }).join('\n') : '';
 }
 
 export default {
-  defaultProps,
-  getGridColumn,
-  getColumnGridColumn,
-  getMediaBreakpointUp,
   makeGridColumns,
 };