+++ src/tests/background-variant.test.js
@@ -1,4 +1,3 @@
-import { fromJS } from 'immutable';
 import { defaultProps, bgVariant } from '../background-variant';
 
 describe('bootstrap background-variant mixins', () => {
@@ -6,10 +5,24 @@ describe('bootstrap background-variant mixins', () => {
     const css = bgVariant(defaultProps['$enable-hover-media-query'], '.toto', '#f00');
     expect(css).not.toContain('undefined');
     expect(css).not.toContain('null');
-    expect(fromJS({ css }).hashCode()).toEqual(-538881241);
+    expect(css).toContain('.toto');
+    expect(css).toContain('background-color: #f00 !important;');
+    expect(css).toContain('a.toto');
+    expect(css).toContain('&:focus,');
+    expect(css).toContain('&.focus,');
+    expect(css).toContain('&:hover,');
+    expect(css).toContain('&.hover');
+    expect(css).toContain('background-color: rgb(204, 0, 0) !important;');
   });
-  it('bgVariant should have paramters', () => {
-    const css = bgVariant();
-    expect(fromJS({ css }).hashCode()).toEqual(-337543508);
+  it('bgVariant should have parameters', () => {
+    const css = bgVariant(!defaultProps['$enable-hover-media-query'], undefined, undefined);
+    expect(css).toContain('undefined');
+    expect(css).toContain('background-color: undefined !important;');
+    expect(css).toContain('aundefined');
+    expect(css).toContain('&:focus,');
+    expect(css).toContain('&.focus');
+    expect(css).toContain('&:hover,');
+    expect(css).toContain('&.hover');
+    expect(css).toContain('background-color: rgb(0, 0, 0) !important;');
   });
 });