+++ src/tests/navbar-toggleable.test.js
@@ -2,7 +2,7 @@ import { fromJS } from 'immutable';
 import { defaultProps, navbarToggleable } from '../navbar-toggleable';
 
 describe('bootstrap toggleable mixins', () => {
-  it('navbarToggleable should return a css without grtid-breakpoints', () => {
+  it('navbarToggleable should return a css without grid-breakpoints', () => {
     const css = navbarToggleable(!defaultProps['$grid-breakpoints']);
     expect(css).not.toContain('undefined');
     expect(css).not.toContain('null');
@@ -12,10 +12,10 @@ describe('bootstrap toggleable mixins', () => {
     const css = navbarToggleable(defaultProps['$grid-breakpoints']);
     expect(css).not.toContain('undefined');
     expect(css).not.toContain('null');
-    expect(fromJS({ css }).hashCode()).toEqual(-767429231);
+    expect(fromJS({ css }).hashCode()).toEqual(-839184507);
   });
   it('navbarToggleable should have arguments', () => {
     const css = navbarToggleable();
-    expect(fromJS({ css }).hashCode()).toEqual(-767429231);
+    expect(fromJS({ css }).hashCode()).toEqual(-839184507);
   });
 });